restructured our directory structure to this:
management_app/
	database/
		dao/
		models/
	play/
		actions/
		controllers/
		requests/
	services/
		authentication/
		common
