+++ play/src/main/scala/com/managementapp/database/dao/UsersAsyncDAO.scala
@@ -1,12 +1,12 @@
-package com.managementapp.authentication.dao
+package com.managementapp.database.dao
 
 import com.google.inject.Inject
-import com.managementapp.authentication.models.{User, UserTable}
-import com.managementapp.common.util.ManagementAppDatabase
+import com.managementapp.database.models.{User, UserTable}
+import com.managementapp.services.common.ManagementAppDatabase
+import play.api.db.slick.DatabaseConfigProvider
 import slick.driver.MySQLDriver.api._
 
 import scala.concurrent.Future
-import play.api.db.slick.DatabaseConfigProvider
 
 
 class UsersAsyncDAO @Inject()(val configProvider: DatabaseConfigProvider) extends ManagementAppDatabase {
