+++ play/src/main/scala/com/managementapp/play/controllers/QuizController.scala
@@ -1,0 +1,5 @@
+package com.managementapp.play.controllers
+
+class QuizController {
+
+}
