--- play/src/main/scala/com/managementapp/profile/controllers/ProfileController.scala
@@ -1,5 +1,0 @@
-package com.managementapp.profile.controllers
-
-class ProfileController {
-
-}
