--- play/src/main/scala/com/managementapp/quiz/controllers/QuizController.scala
@@ -1,5 +1,0 @@
-package com.managementapp.quiz.controllers
-
-class QuizController {
-
-}
